+++ src/debug-server.ts
@@ -25,7 +25,6 @@ import { randomUUID } from "crypto";
 import {
   AnySchema,
   SchemaOutput,
-  ZodRawShapeCompat,
 } from "@modelcontextprotocol/sdk/server/zod-compat.js";
 import { ToolConfig } from "./tools/types";
 import { logger } from "./logger";
@@ -156,15 +155,15 @@ export class DebugServer extends EventEmitter implements DebugServerEvents {
       server.registerTool(config.name, config, toolWrapped);
     }
 
-    registerToolWrapper(setBreakpoint.tool, handleSetBreakpoint);
-    registerToolWrapper(removeBreakpoint.tool, handleRemoveBreakpoint);
-    registerToolWrapper(variables.tool, handleVariables);
-    registerToolWrapper(evaluate.tool, handleEvaluate);
-    registerToolWrapper(launch.tool, handleLaunch);
-    registerToolWrapper(stop.tool, handleStop);
-    registerToolWrapper(cont.tool, handleContinue);
-    registerToolWrapper(waitForBreakpoint.tool, handleWaitForBreakpoint);
-    registerToolWrapper(getWorkspace.tool, handleGetWorkspace);
+    registerToolWrapper(setBreakpoint.tool, setBreakpoint.handle);
+    registerToolWrapper(removeBreakpoint.tool, removeBreakpoint.handle);
+    registerToolWrapper(variables.tool, variables.handle);
+    registerToolWrapper(evaluate.tool, evaluate.handle);
+    registerToolWrapper(launch.tool, launch.handle);
+    registerToolWrapper(stop.tool, stop.handle);
+    registerToolWrapper(cont.tool, cont.handle);
+    registerToolWrapper(waitForBreakpoint.tool, waitForBreakpoint.handle);
+    registerToolWrapper(getWorkspace.tool, getWorkspace.handle);
   }
 
   stop(): Promise<void> {
@@ -180,261 +179,5 @@ export class DebugServer extends EventEmitter implements DebugServerEvents {
         resolve();
       });
     });
-  }
-}
-
-function cleanStackFrames(stackFrames_: any) {
-  const stackFramesSchema = z.array(
-    z.object({
-      line: z.number(),
-      source: z
-        .object({
-          name: z.string().optional(),
-          path: z.string().optional(),
-        })
-        .loose()
-        .optional(),
-      id: z.number(),
-      name: z.string(),
-      column: z.number(),
-      presentationHint: z.string().optional(),
-    }),
-  );
-
-  const stackFrames = stackFramesSchema.parse(stackFrames_);
-
-  const stackFramesTransformed = stackFrames.map((frame) => {
-    if (!frame.source?.path) {
-      return frame;
-    }
-    let pathURL;
-    try {
-      pathURL = new URL(frame.source.path);
-    } catch (e) {
-      return frame;
-    }
-    return {
-      ...frame,
-      source: {
-        ...frame.source,
-        path:
-          pathURL.protocol === "vscode-remote:"
-            ? pathURL.pathname
-            : pathURL.toString(),
-      },
-    };
-  });
-
-  return stackFramesTransformed;
-}
-
-function formatStackFrames(
-  stackFrames: ReturnType<typeof cleanStackFrames>,
-): string {
-  // Collapse internal frames
-  const res = stackFrames.reduce(
-    ({ internalFramesCounter, acc }, frame) => {
-      if (frame.presentationHint === "subtle") {
-        return { internalFramesCounter: internalFramesCounter + 1, acc };
-      } else {
-        return {
-          internalFramesCounter: 0,
-          acc: `${acc}${internalFramesCounter > 0 ? `... (${internalFramesCounter} internal frames)\n` : ""}${JSON.stringify(frame)}\n`,
-        };
-      }
-    },
-    { internalFramesCounter: 0, acc: "" },
-  );
-  // Handle trailing internal frames
-  return (
-    res.acc +
-    (res.internalFramesCounter > 0
-      ? `... (${res.internalFramesCounter} internal frames)\n`
-      : "")
-  );
-}
-
-function handleGetWorkspace(): string {
-  const workspaceFolder = vscode.workspace.workspaceFolders?.[0];
-  if (!workspaceFolder) {
-    return "No workspace folder found";
-  }
-  return workspaceFolder.uri.fsPath;
-}
-
-async function handleLaunch(): Promise<string> {
-  const workspaceFolder = vscode.workspace.workspaceFolders?.[0];
-  if (!workspaceFolder) {
-    throw new Error("No workspace folder found");
-  }
-
-  // Return an error message if we are already debugging
-  if (vscode.debug.activeDebugSession) {
-    return "Already debugging.";
-  }
-
-  // Start debugging using the well-known launch configuration
-  await vscode.debug.startDebugging(workspaceFolder, "claude_debug");
-
-  return `Launched.`;
-}
-
-async function handleStop(): Promise<string> {
-  await vscode.debug.stopDebugging();
-
-  return `Stopped debugging.`;
-}
-
-async function handleWaitForBreakpoint(): Promise<string> {
-  const frame = await waitForStackFrame();
-  if (!frame) {
-    return "Debuggee has exited.";
-  }
-  const { session, threadId } = frame;
-
-  const stack = await session.customRequest("stackTrace", { threadId });
-  return `Stopped:\n${formatStackFrames(cleanStackFrames(stack.stackFrames))}`;
-}
-
-// Source - https://stackoverflow.com/a/46675277
-// Posted by Drew Noakes, modified by community. See post 'Timeline' for change history
-// Retrieved 2026-03-28, License - CC BY-SA 4.0
-function withTimeout<T>(millis: number, promise: Promise<T>): Promise<T> {
-  let timeout: NodeJS.Timeout;
-  const timeoutPromise = new Promise<T>(
-    (resolve, reject) =>
-      (timeout = setTimeout(
-        () => reject(`Timed out after ${millis} ms.`),
-        millis,
-      )),
-  );
-  return Promise.race([promise, timeoutPromise]).finally(() => {
-    if (timeout) {
-      clearTimeout(timeout);
-    }
-  });
-}
-
-async function waitForStackFrame(): Promise<
-  vscode.DebugStackFrame | undefined
-> {
-  if (!vscode.debug.activeDebugSession) {
-    return undefined;
-  }
-
-  const stackItem = vscode.debug.activeStackItem;
-  if (stackItem instanceof vscode.DebugStackFrame) {
-    return stackItem;
-  }
-
-  let handle: vscode.Disposable;
-  const sfPromise = new Promise<vscode.DebugStackFrame | undefined>((res) => {
-    handle = vscode.debug.onDidChangeActiveStackItem((stackItem) => {
-      if (stackItem instanceof vscode.DebugStackFrame) {
-        res(stackItem);
-      }
-      if (typeof stackItem === "undefined") {
-        res(undefined);
-      }
-    });
-  });
-  return withTimeout(15000, sfPromise).finally(() => handle?.dispose());
-}
-
-async function handleContinue() {
-  let session = vscode.debug.activeDebugSession;
-  if (!session) {
-    throw new Error("No active debug session");
-  }
-
-  // Get the current thread ID (required by DAP spec)
-  const threads = await session.customRequest("threads");
-  let threadId = threads.threads[0].id;
-
-  // Continue with the thread ID
-  await session.customRequest("continue", { threadId });
-
-  return `Continued.`;
-}
-
-async function handleSetBreakpoint(
-  payload: z.infer<typeof setBreakpoint.tool.inputSchema>,
-) {
-  // Open the file and make it active
-  const document = await vscode.workspace.openTextDocument(payload.file);
-  const editor = await vscode.window.showTextDocument(document);
-
-  const bp = new vscode.SourceBreakpoint(
-    new vscode.Location(
-      editor.document.uri,
-      new vscode.Position(payload.line - 1, 0),
-    ),
-    true,
-    payload.condition,
-  );
-  vscode.debug.addBreakpoints([bp]);
-  return "Success.";
-}
-
-async function handleRemoveBreakpoint(
-  payload: z.infer<typeof removeBreakpoint.tool.inputSchema>,
-) {
-  const bps = vscode.debug.breakpoints.filter((bp) => {
-    if (bp instanceof vscode.SourceBreakpoint) {
-      return bp.location.range.start.line === payload.line - 1;
-    }
-    return false;
-  });
-  if (bps.length === 0) {
-    return `No breakpoint on line ${payload.line}`;
-  }
-  vscode.debug.removeBreakpoints(bps);
-  return "Success.";
-}
-
-async function handleVariables(
-  payload: z.infer<typeof variables.tool.inputSchema>,
-) {
-  const session = vscode.debug.activeDebugSession;
-  if (!session) {
-    return "No active debug session.";
-  }
-
-  const response = await session.customRequest("variables", {
-    variablesReference: payload.variablesReference,
-  });
-
-  return `Variables result: ${JSON.stringify(response)}`;
-}
-
-async function handleEvaluate(
-  payload: z.infer<typeof evaluate.tool.inputSchema>,
-) {
-  const session = vscode.debug.activeDebugSession;
-  if (!session) {
-    return "No active debug session.";
-  }
-
-  try {
-    const response = await session.customRequest("evaluate", {
-      expression: payload.expression,
-      frameId: payload.frameId,
-      context: "watch",
-    });
-
-    return `Eval result: ${JSON.stringify(response)}`;
-  } catch (err: any) {
-    let errorMessage = "";
-    let stackTrace = "";
-
-    if (err instanceof Error) {
-      errorMessage = err.message;
-      if (err.stack) {
-        stackTrace = `\nStack: ${err.stack}`;
-      }
-    } else {
-      errorMessage = String(err);
-    }
-    return `Eval failed: ${errorMessage}${stackTrace}`;
   }
 }
